muchas cosas

+++ src/modelo/ModuloObjetivo/BajarPeso.java
@@ -1,10 +1,10 @@
 package modelo.moduloObjetivo;
 
 public class BajarPeso extends ObjetivoStrategy {
-	private float pesoInicial;
-	private float alturaInicial;
+	private double pesoInicial;
+	private double alturaInicial;
 
-	public BajarPeso(double duracion,  float peso, float altura) {
+	public BajarPeso(double duracion,  double peso, double altura) {
 		super(duracion, 0, 3);
 		this.pesoInicial = peso;
 		this.alturaInicial = altura;
@@ -31,11 +31,11 @@ public class BajarPeso extends ObjetivoStrategy {
 		return "BajarPeso";
 	}
 
-	public float getAlturaInicial(){
+	public double getAlturaInicial(){
 		return this.alturaInicial;
 	}
 
-	public float getPesoInicial() {
+	public double getPesoInicial() {
 		return this.pesoInicial;
 	}
